backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
google-genai==0.1.0
PyPDF2==3.0.1
pandas==2.1.3
openpyxl==3.1.2
httpx==0.25.1
pydantic==2.5.0
websockets>=13.0
pdf2image==1.16.3
Pillow==10.1.0
requests==2.31.0
aiohttp==3.9.1
aiofiles==23.2.1
reportlab==4.0.6